change the type data of membershipNumber

+++ src/main/java/com/example/demo/entity/Member.java
@@ -24,7 +24,7 @@ import java.util.Locale;
 public class Member {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer membershipNumber;
+    private Long membershipNumber;
 
     private String firstName;
